try to add crosshair, not fixed

+++ src/Managers/CrossManager.java
@@ -5,9 +5,13 @@ import org.academiadecodigo.simplegraphics.pictures.Picture;
 public class CrossManager {
 
     private Picture sprite;
+    private boolean isVisible = false;
 
     public CrossManager() {
+        createSprite();
+    }
 
+    private void createSprite() {
         this.sprite = new Picture(0, 0, "Resources/CROSSHAIR.png");
 
         int size = 40;
@@ -16,9 +20,14 @@ public class CrossManager {
         sprite.grow(dw, dh);
 
         sprite.draw();
+        isVisible = true;
     }
 
     public void updatePosition(double mouseX, double mouseY) {
+        // Se a crosshair foi deletada, recriá-la
+        if (!isVisible) {
+            createSprite();
+        }
 
         double x = mouseX - sprite.getWidth() / 2.0;
         double y = mouseY - sprite.getHeight() / 2.0;
@@ -26,7 +35,21 @@ public class CrossManager {
         sprite.translate(x - sprite.getX(), y - sprite.getY());
     }
 
+    public void show() {
+        if (!isVisible) {
+            createSprite();
+        }
+    }
+
     public void hide() {
-        sprite.delete();
+        if (isVisible && sprite != null) {
+            sprite.delete();
+            isVisible = false;
+        }
+    }
+
+    public void destroy() {
+        hide();
+        sprite = null;
     }
 }
